pyproject.toml
[tool.poetry]
name = "rbb-bot"
version = "2.0.0"
description = ""
authors = ["badmagick329"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.0"
"discord.py" = "^2.1"
aiohttp = "^3.8.3"
yarl = "^1.8.1"
aioscheduler = "^1.4.2"
asyncpg = "^0.26.0"
pydantic = "^1.10.2"
numpy = "^1.23.3"
pendulum = "^2.1.2"
peony-twitter = {extras = ["all"], version = "^2.1.2"}
Pillow = "^9.2.0"
psycopg2-binary = "^2.9.3"
PyYAML = "^6.0"
beautifulsoup4 = "^4.11.1"
tortoise-orm = "^0.19.2"
aerich = "^0.7.1"
asyncpraw = "^7.6.0"
lxml = "^4.9.1"
python-rapidjson = "^1.8"
dateparser = "^1.1.2"
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus"}
jishaku = "^2.5.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
black = "^22.10.0"
ipykernel = "^6.16.0"
pytest-asyncio = "^0.19.0"
snakeviz = "^2.1.1"


[tool.aerich]
tortoise_orm = "dbconfig.DB_CONFIG"
location = "./migrations/"
src_folder = "./rbb_bot"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-rP"
testpaths = [
    "tests",
]